media-app/backend/requirements.txt
alembic==1.13.2
argcomplete==3.5.0
awscli==1.34.13
bidict==0.23.1
blinker==1.8.2
boto3==1.35.13
botocore==1.35.13
cachelib==0.13.0
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.16
durationpy==0.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.4.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
graphviz==0.20.3
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
Mako==1.3.5
MarkupSafe==2.1.5
msgspec==0.18.6
mypy==1.11.2
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
packaging==24.1
placebo==0.9.0
pyasn1==0.6.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
setuptools==74.1.2
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.34
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
wsproto==1.2.0
WTForms==3.1.2
zappa==0.59.0
